% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngscleanR-package.R
\docType{package}
\name{ngscleanR-package}
\alias{ngscleanR}
\alias{ngscleanR-package}
\title{ngscleanR: Helper Functions for Cleaning Player Tracking Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
