% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_play}
\alias{plot_play}
\title{Plot a play lol}
\usage{
plot_play(
  df_track,
  orientation = TRUE,
  dot_size = 6,
  segment_length = 2.5,
  segment_size = 1.5,
  numbers = TRUE,
  animated = TRUE,
  animated_h = 4,
  animated_w = 8,
  animated_res = 200,
  frame = NULL
)
}
\arguments{
\item{df_track}{A df of tracking data}

\item{orientation}{Show lines representing where player is facing (default = T)}

\item{dot_size}{Size of player dots (default = 6)}

\item{segment_length}{Length of orientation segment lines (default = 2.5)}

\item{segment_size}{Width of orientation segment lines (default = 1.5)}

\item{numbers}{Show player jersey numbers (default = T)}

\item{animated}{Whether play is animated or a still frame (default = T)}

\item{animated_h}{If animated, height of animated image (default = 4)}

\item{animated_w}{If animated, width of animated image (default = 8)}

\item{animated_res}{If animated, resolution of animated image (default = 200)}

\item{frame}{frame_id to plot (default = NULL, ie plot all provided frames).}
}
\description{
Plot a play lol
}
