% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{compute_o_diff}
\alias{compute_o_diff}
\title{Compute orientation difference}
\usage{
compute_o_diff(df, prefix = "qb")
}
\arguments{
\item{df}{A dataframe containing x, y, o, "prefix"_x, and "prefix"_y}

\item{prefix}{(default = "qb"). Columns prefix_x and prefix_y must be contained in \code{df}. These columns
contain the x and y locations that will be used to calculate orientation difference.}
}
\value{
Original dataframe with o_to_"prefix" added, which is the difference in orientation
in degrees between the way the player is facing and where the "prefix" player is (0 is facing
directly at the "prefix" player, 180 is directly away).
}
\description{
Compute difference in orientation between direction player is currently facing and
orientation if player were facing towards a given x and y location.
}
\examples{
df <- tibble::tibble("x" = 20, "y" = 30, "o" = 270, "qb_x" = 10, "qb_y" = 25)
df <- compute_o_diff(df)
str(df)
}
