% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{cut_plays}
\alias{cut_plays}
\title{Trim plays based on events}
\usage{
cut_plays(
  df,
  end_events = c("pass_forward", "qb_sack", "qb_strip_sack", "qb_spike", "tackle",
    "pass_shovel"),
  time_after_event = 0,
  throw_frame = 25
)
}
\arguments{
\item{df}{A dataframe}

\item{end_events}{(default c("pass_forward", "qb_sack", "qb_strip_sack", "qb_spike", "tackle", "pass_shovel"))
Events designated as play end events}

\item{time_after_event}{Number of frames to keep after the event (default: 0). There are 10 frames in each second.}

\item{throw_frame}{If the end event happens before this frame, remove this play from the df (default: 25, ie 1.5 seconds
into the play). To not employ play dropping, provide throw_frame = NULL.}
}
\value{
The original df with trimmed frames (and if throw_frame provided, short plays removed).
}
\description{
Trim plays based on events
}
